implementing protobufs
created protobuf branch

created template to make appending to protobufs easier for users. The
program should be able to read it, and convert it to a temporary file
that will then be serialized and transferred into a protobuf. Still
working on implementing this correctly.

changed a few other things as well. Reference the commit history for
other changes.