created branch on remote repo
made some changes to reference.go, see commit history.

the program works as intended up to opening the temporary template.
after which, the program continues on without waiting for changes to be
written to the temporary file. I need to learn more about goroutines in
order to understand how to fix the issue properly.