partial-fix: reference command recognizes neovim
I was able to get the behaviour I was looking for with the reference
command -a operation. It will permit opening and writing to the temporary file
with neovim properly, prior to continuing with the rest of the program.
I still need to work on getting it working with external text editors
like kate and vscode. However, this is good enough for now and will
allow me to move on to working on the other functions of the command.