feat: list flag added, write to protobuf implemented
after successfully implementing these two, it became apparent that I
need to append the new data to an existing reference, which is a bit
more involved than I anticipated. That will be worked on and implemented
in the next commit. I'm just happy that I can write to the protobuf file
now. will also work on writing to the "correct" location.

also: refactored reference.proto. Regenerated the reference.pb.go file
as well.